+++ scripts/preset-js.js
@@ -1,10 +1,15 @@
+const presetEnv = {
+  modules: process.env.RUN_MODE === "es" ? false : "commonjs"
+};
+
+if (process.env.RUNTIME_ENV === "server") {
+  presetEnv.targets = { node: "current" };
+} else {
+  presetEnv.targets = { browsers: ["last 2 versions", "safari >= 7"] };
+}
+
 module.exports = {
   presets: [
-    ['babel-preset-env', {
-      targets: {
-        browsers: ['last 2 versions', 'safari >= 7']
-      },
-      modules: process.env.RUN_MODE === 'es' ? false : 'commonjs'
-    }],
+    ["babel-preset-env", presetEnv],
   ],
 };